Resolve code review comments.
