+++ packages/fast-components-styles-msft/src/radio/index.ts
@@ -1,8 +1,4 @@
-import {
-    applyDisabledState,
-    DesignSystem,
-    withDesignSystemDefaults,
-} from "../design-system";
+import { DesignSystem, withDesignSystemDefaults } from "../design-system";
 import {
     ComponentStyles,
     ComponentStyleSheet,
@@ -26,6 +22,7 @@ import {
     neutralOutlineRest,
 } from "../utilities/color";
 import { applyFontSize, densityToTypeOffset, padding } from "../utilities/density";
+import { applyDisabledState } from "../utilities/disabled";
 
 const styles: ComponentStyles<RadioClassNameContract, DesignSystem> = (
     config: DesignSystem
