+++ packages/fast-components-styles-msft/src/text-field/index.ts
@@ -1,6 +1,5 @@
 import { DesignSystem, withDesignSystemDefaults } from "../design-system";
 import { TextFieldClassNameContract } from "@microsoft/fast-components-class-name-contracts-base";
-import { toPx } from "@microsoft/fast-jss-utilities";
 import { ComponentStyles, ComponentStyleSheet } from "@microsoft/fast-jss-manager";
 import { height, maxHeight, minHeight } from "../utilities/density";
 import { inputFieldStyles } from "../patterns/input-field";
